Restructure tests into multiple files.
